make py scripts more modular
